apps/api/requirements.txt
# FastAPI e servidor ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Banco de dados (async)
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Validação e configuração
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Autenticação
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Cache e filas
redis==5.0.1
hiredis==2.3.2

# Testes
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilitários
python-dotenv==1.0.0

# Extens�es Geo/Vector
geoalchemy2==0.14.3
pgvector==0.2.4
